pyproject.toml
[project]
name = "food-data-app"
version = "0.1.0"
description = "A CRUD web application built with Python Shiny and DuckDB for managing food consumption data."
authors = [
 {name = "James Westwood", email = "james.westwood@food.gov.uk"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
]

[tool.poetry]
packages = [
 { include = "**/*.py" }
]

[tool.poetry.dependencies]
python = "^3.9"
shiny = "^0.6.0"
shinyswatch = "^0.4.0"
duckdb = "^0.9.0"
pandas = "^2.0.0"
xlsxwriter = "^3.1.0"
openpyxl = "^3.1.0"
pyyaml = "^6.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.5"
mkdocstrings = {extras = ["python"], version = "^0.28.1"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"